修复 hlsl 类型转换

+++ modules/gui/renderer/shaders/pixel.ps_6_0.hlsl
@@ -14,8 +14,8 @@ struct VSOut
 void main(VSOut input,     
     out float4 o_color : SV_Target0) : SV_TARGET
 {
-    const float visible = all(input.clip_uv == clamp(input.clip_uv, float2(-1.f, -1.f), float2(1.f, 1.f)));
-	const float visible2 = all(input.clip_uv2 == clamp(input.clip_uv2, float2(-1.f, -1.f), float2(1.f, 1.f)));
+    const float visible = (float)all(input.clip_uv == clamp(input.clip_uv, float2(-1.f, -1.f), float2(1.f, 1.f)));
+	const float visible2 = (float)all(input.clip_uv2 == clamp(input.clip_uv2, float2(-1.f, -1.f), float2(1.f, 1.f)));
 	const float edge = min(1.0, (1.0-abs((input.aa.x * 2.0) - 1.0)) * input.aa.y);
 	o_color = input.color;
 	o_color.w = o_color.w * visible * visible2 * edge;
